+++ ultimate/testData/spring/core/rename/factoryMethodParamByXmlRef/after/test.kt
@@ -1,0 +1,6 @@
+package a
+
+class Construction(param: Int)
+object FactoryObject {
+    @JvmStatic fun buildObject(param2: Int) = Construction(param2)
+}